Implement lazy parsing mode.
Lazy mode now just consists of a special case in the player loop:
if there is some data left to process, read it for as long as
no key was pressed and the frame delay hasn't expired.

Once no data is left to process, delete all traces of the reader
and let the player proceed in eager processing mode.