Partially fix bug; rebrand it as a feature.
Jumping to end of currently available stream would stop waiting for
frame delays altogether.

The reason for this is that the frame delay is read from parrays as
the difference between the current frame's timestamp and the next
one's.  But when the next one isn't available, parray_search
yields the current one anyway, hence a difference of 0.

This is actually behavior that could be wanted--"read ttyrec as
fast as processed."  So I changed the G command to stop one frame
before the last one in lazy mode, and added F to read as fast as
possible.  F as in less's follow mode.