Clarify the situation around the end of the film.
Disallow seeking frames past the end of the film in eager mode.
Seeking a frame past the end of the film in lazy mode while
playing goes there as fast as possible, then keeps playing
normally.  If that frame turns out not existing, the counter clips
to the last one that does as soon as the information is available.